extracted lists of tasks from Tasks components and created new TaskList component
